anki_for_jurafsky_slp3/pyproject.toml
[project]
name = "anki-for-jurafsky-slp3"
version = "0.1.0"
description = "Anki flashcard generator for Jurafsky & Martin Speech and Language Processing (3rd ed)"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Krzysztof Jurkiewicz"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "diskcache>=5.6.3",
    "genanki>=0.13.1",
    "google-genai>=1.25.0",
    "openai>=2.3.0",
    "pandas>=2.3.3",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.ruff.lint]
ignore = ["E741"]
